+++ src/main/java/controller/AbstractController.java
@@ -7,11 +7,11 @@ import java.io.IOException;
 
 public abstract class AbstractController implements Controller {
 
-    void doGet(final HttpRequest request, final HttpResponse response) throws IOException {
+    protected void doGet(final HttpRequest request, final HttpResponse response) throws IOException {
         response.makeResponse();
     }
 
-    void doPost(final HttpRequest request, final HttpResponse response) throws IOException {
+    protected void doPost(final HttpRequest request, final HttpResponse response) throws IOException {
         response.makeResponse();
     }
 }
